+++ app/src/main/java/ar/scacchipa/e_commerce/app/DetailFragment.kt
@@ -1,22 +1,22 @@
-package ar.scacchipa.e_commerce
+package ar.scacchipa.e_commerce.app
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import ar.scacchipa.e_commerce.databinding.CartScreenBinding
+import ar.scacchipa.e_commerce.databinding.DetailScreenBinding
 
-class CartFragment: Fragment() {
+class DetailFragment: Fragment() {
 
-    private var binding: CartScreenBinding? = null
+    private var binding: DetailScreenBinding? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = CartScreenBinding.inflate(inflater, container, false)
+        binding = DetailScreenBinding.inflate(inflater, container, false)
         return binding?.root
     }
 }
