+++ app/src/main/java/ar/scacchipa/e_commerce/app/MainActivity.kt
@@ -1,7 +1,8 @@
-package ar.scacchipa.e_commerce
+package ar.scacchipa.e_commerce.app
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import ar.scacchipa.e_commerce.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
