+++ app/src/main/res/layout/highlight_item_layout.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ar.scacchipa.e_commerce.SquareCardView
+<ar.scacchipa.e_commerce.ui.SquareCardView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -57,4 +57,4 @@
             app:layout_constraintGuide_percent="0.75" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
-</ar.scacchipa.e_commerce.SquareCardView>
+</ar.scacchipa.e_commerce.ui.SquareCardView>
